ref: CSV import test; instance offline
